DivOperator fails with InvalidArgumentException
